requirements.txt
discord.py
pytz
Pillow
